forecast-service/requirements.txt
fastapi
uvicorn
prophet
statsmodels
pandas
numpy
supabase
python-dateutil
